SDD should be Engineered to be Progressive
Recommending to add a statement to the **Engineered** section of SDD. A key distinction in modern application development is the ability to separate the act of deployment from the act of release. Ideally this is done in a progressive fashion that allows the owner of the release to limit the impact of changes in the case of canary deploys or percentage rollouts. 

Here are a few resources that speak to this at greater length:

https://redmonk.com/jgovernor/2018/08/06/towards-progressive-delivery/
https://blog.turbinelabs.io/deploy-not-equal-release-part-one-4724bc1e726b
https://launchdarkly.com/blog/progressive-delivery-a-history-condensed/